requirements.txt
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.5.0
Pillow==10.2.0
python-dotenv==1.0.1
scikit_learn==1.4.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
